Fixed dragging issues and revised backend
Dragging plants was finicky because plant colliders were behind the drop
zone colliders relative to the camera position. Fixed this issue by
moving the plants to a z-position of -1 which brought it closer to the
camera.

Plants now must be watered correctly for the number of days they were
unhealthy for in order to become healthy again. The coloring only
applies to the unhealthy states.